Fixed bugs in orders

+++ database/migrations/2019_03_06_140957_create_customers_table.php
@@ -18,7 +18,7 @@ class CreateCustomersTable extends Migration
             $table->string('name')->nullable();
             $table->string('surname')->nullable();
             $table->string('company')->nullable();
-            $table->integer('telephone')->nullable();
+            $table->bigInteger('telephone')->nullable();
             $table->string('email')->nullable();
             $table->integer('orders_sum')->nullable();
             $table->integer('source_id')->nullable();
